refactor: localize permission approval continuations
